requirements.txt
absl-py==2.1.0
altair==5.3.0
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.4.0
catboost==1.2.5
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
docutils==0.21.2
flatbuffers==24.3.25
fonttools==4.53.1
fpdf==1.7.2
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.65.4
h5py==3.11.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.4.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pipdeptree==2.23.1
plotly==5.23.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
smmap==5.0.1
statistics==1.0.3.5
streamlit==1.37.1
tenacity==8.5.0
urllib3==2.2.2
watchdog==4.0.2
Werkzeug==3.0.3
wrapt==1.16.0
xgboost==2.1.1
